+++ crates/runtime/src/execution/datafusion/functions/bool.rs
@@ -6,243 +6,12 @@ use datafusion_expr::{ScalarFunctionArgs, ScalarUDFImpl};
 use half::f16;
 use std::any::Any;
 
-// booland SQL function
-// Computes the Boolean AND of two numeric expressions. In accordance with Boolean semantics:
-// Syntax: BOOLAND( <expr1> , <expr2> )
-// - <expr1>: Non-zero values (including negative numbers) are regarded as True. Zero values are regarded as False.
-// - <expr2>: Non-zero values (including negative numbers) are regarded as True. Zero values are regarded as False.
-// Note: `booland` returns
-// - True if both expressions are non-zero.
-// - False if both expressions are zero or one expression is zero and the other expression is non-zero or NULL.
-// - NULL if both expressions are NULL or one expression is NULL and the other expression is non-zero.
-#[derive(Debug)]
-pub struct BoolAndFunc {
-    signature: Signature,
-}
 
-impl Default for BoolAndFunc {
-    fn default() -> Self {
-        Self::new()
-    }
-}
 
-impl BoolAndFunc {
-    pub fn new() -> Self {
-        Self {
-            signature: Signature::comparable(2, Volatility::Immutable),
-        }
-    }
-}
 
-impl ScalarUDFImpl for BoolAndFunc {
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
 
-    fn name(&self) -> &'static str {
-        "booland"
-    }
 
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
 
-    fn return_type(&self, _arg_types: &[DataType]) -> DFResult<DataType> {
-        Ok(DataType::Boolean)
-    }
-
-    fn invoke_with_args(&self, args: ScalarFunctionArgs) -> DFResult<ColumnarValue> {
-        let lhs = match &args.args[0] {
-            ColumnarValue::Scalar(val) => val.to_owned(),
-            ColumnarValue::Array(array) => ScalarValue::try_from_array(&array, 0)?,
-        };
-
-        let rhs = match &args.args[1] {
-            ColumnarValue::Scalar(val) => val.to_owned(),
-            ColumnarValue::Array(array) => ScalarValue::try_from_array(&array, 0)?,
-        };
-
-        if lhs.is_null() && rhs.is_null() {
-            return Ok(ColumnarValue::Scalar(ScalarValue::Boolean(None)));
-        }
-
-        if (lhs.is_null() || rhs.is_null()) && (is_true(&lhs)? || is_true(&rhs)?) {
-            return Ok(ColumnarValue::Scalar(ScalarValue::Boolean(None)));
-        }
-        Ok(ColumnarValue::Scalar(ScalarValue::from(
-            is_true(&lhs)? && is_true(&rhs)?,
-        )))
-    }
-}
-
-// boolor SQL function
-// Computes the Boolean OR of two numeric expressions. In accordance with Boolean semantics:
-// - Non-zero values (including negative numbers) are regarded as True.
-// - Zero values are regarded as False.
-// Syntax: BOOLOR( <expr1> , <expr2> )
-// Note: `boolor` returns
-// - True if both expressions are non-zero or one expression is non-zero and the other expression is zero or NULL.
-// - False if both expressions are zero.
-// - NULL if both expressions are NULL or one expression is NULL and the other expression is zero.
-#[derive(Debug)]
-pub struct BoolOrFunc {
-    signature: Signature,
-}
-
-impl Default for BoolOrFunc {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl BoolOrFunc {
-    pub fn new() -> Self {
-        Self {
-            signature: Signature::comparable(2, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for BoolOrFunc {
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-
-    fn name(&self) -> &'static str {
-        "boolor"
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, _arg_types: &[DataType]) -> DFResult<DataType> {
-        Ok(DataType::Boolean)
-    }
-
-    fn invoke_with_args(&self, args: ScalarFunctionArgs) -> DFResult<ColumnarValue> {
-        let lhs = match &args.args[0] {
-            ColumnarValue::Scalar(val) => val.to_owned(),
-            ColumnarValue::Array(array) => ScalarValue::try_from_array(&array, 0)?,
-        };
-
-        let rhs = match &args.args[1] {
-            ColumnarValue::Scalar(val) => val.to_owned(),
-            ColumnarValue::Array(array) => ScalarValue::try_from_array(&array, 0)?,
-        };
-
-        if lhs.is_null() && rhs.is_null() {
-            return Ok(ColumnarValue::Scalar(ScalarValue::Boolean(None)));
-        }
-
-        if (lhs.is_null() || rhs.is_null()) && (!is_true(&lhs)? && !is_true(&rhs)?) {
-            return Ok(ColumnarValue::Scalar(ScalarValue::Boolean(None)));
-        }
-        Ok(ColumnarValue::Scalar(ScalarValue::from(
-            is_true(&lhs)? || is_true(&rhs)?,
-        )))
-    }
-}
-
-// boolxor SQL function
-// Computes the Boolean XOR of two numeric expressions (i.e. one of the expressions, but not both expressions, is TRUE). In accordance with Boolean semantics:
-// - Non-zero values (including negative numbers) are regarded as True.
-// - Zero values are regarded as False.
-// Syntax: BOOLXOR( <expr1> , <expr2> )
-// Note: `boolxor` returns
-// - True if one expression is non-zero and the other expression is zero.
-// - False if both expressions are non-zero or both expressions are zero.
-// - NULL if one or both expressions are NULL.
-#[derive(Debug)]
-pub struct BoolXorFunc {
-    signature: Signature,
-}
-
-impl Default for BoolXorFunc {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl BoolXorFunc {
-    pub fn new() -> Self {
-        Self {
-            signature: Signature::comparable(2, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for BoolXorFunc {
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-
-    fn name(&self) -> &'static str {
-        "boolxor"
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, _arg_types: &[DataType]) -> DFResult<DataType> {
-        Ok(DataType::Boolean)
-    }
-
-    fn invoke_with_args(&self, args: ScalarFunctionArgs) -> DFResult<ColumnarValue> {
-        let lhs = match &args.args[0] {
-            ColumnarValue::Scalar(val) => val.to_owned(),
-            ColumnarValue::Array(array) => ScalarValue::try_from_array(&array, 0)?,
-        };
-
-        let rhs = match &args.args[1] {
-            ColumnarValue::Scalar(val) => val.to_owned(),
-            ColumnarValue::Array(array) => ScalarValue::try_from_array(&array, 0)?,
-        };
-
-        if lhs.is_null() && rhs.is_null() {
-            return Ok(ColumnarValue::Scalar(ScalarValue::Boolean(None)));
-        }
-
-        if lhs.is_null() || rhs.is_null() {
-            return Ok(ColumnarValue::Scalar(ScalarValue::Boolean(None)));
-        }
-        if (is_true(&lhs)? && !is_true(&rhs)?) || (!is_true(&lhs)? && is_true(&rhs)?) {
-            Ok(ColumnarValue::Scalar(ScalarValue::from(Some(true))))
-        } else if (is_true(&lhs)? && is_true(&rhs)?) || (!is_true(&lhs)? && !is_true(&rhs)?) {
-            Ok(ColumnarValue::Scalar(ScalarValue::from(Some(false))))
-        } else {
-            Ok(ColumnarValue::Scalar(ScalarValue::Boolean(None)))
-        }
-    }
-}
-
-fn is_true(v: &ScalarValue) -> DFResult<bool> {
-    if v.is_null() {
-        return Ok(false);
-    }
-    Ok(match v {
-        ScalarValue::Float16(Some(v)) => *v != f16::from_f32(0.0),
-        ScalarValue::Float32(Some(v)) => *v != 0.,
-        ScalarValue::Float64(Some(v)) => *v != 0.,
-        ScalarValue::Decimal128(Some(v), _, _) => *v != 0,
-        ScalarValue::Decimal256(Some(v), _, _) => *v != i256::from_i128(0),
-        ScalarValue::Int8(Some(v)) => *v != 0,
-        ScalarValue::Int16(Some(v)) => *v != 0,
-        ScalarValue::Int32(Some(v)) => *v != 0,
-        ScalarValue::Int64(Some(v)) => *v != 0,
-        ScalarValue::UInt8(Some(v)) => *v != 0,
-        ScalarValue::UInt16(Some(v)) => *v != 0,
-        ScalarValue::UInt32(Some(v)) => *v != 0,
-        ScalarValue::UInt64(Some(v)) => *v != 0,
-        _ => {
-            return Err(datafusion_common::DataFusionError::Internal(format!(
-                "Unsupported type {:?}",
-                v.data_type()
-            )));
-        }
-    })
-}
 
 #[cfg(test)]
 mod tests {
